requirements.txt
aiogram==3.26.0
matplotlib==3.10.1
openpyxl==3.1.5
aiohttp==3.11.12
pytz==2025.2
python-dotenv==1.0.0
feedparser==6.0.11
apscheduler==3.10.4
flask==3.1.2
deep-translator>=1.11.4